Split Prints into its own component and imported into the prints page
