+++ src/data/prints.json
@@ -1,9 +1,9 @@
 [
   {
-    "name": "A member of the Adirondack Motor Enthusiasts Club races their Honda Civic across a frozen Lake George at the Lake George Winter Carnival",
-    "description": "Lake George, NY",
+    "name": "Superwolves: Will Oldham & Matt Sweeney",
+    "description": "Print size: 9x6 | Canvas size: 10x7",
     "category": "prints",
-    "src": "../assets/images/action/lake-george-ice-racing-20190223-_MG_1645.jpg",
+    "src": "../assets/images/musicians/superwolves-7752.jpg",
     "price": 125.0
   }
 ]
